# swarm/network: fix data race in fetcher_test.go