**Problem**
Let's say TestA() and TestB() was started right after each other.
TestA() called `go Fetcher.run(ctx)` (read: searchTimeout), on test
completion the context was cancelled, but the test did not wait for
goroutine termination. Test(B) started by modifying searchTimeout
(write).

**Solution**
Let's just store searchTimeout on the Fetcher struct and change the package var (default) to const (thread-safe).

**Alternative (almost) solution**
The above could have been solved by (little ugly, more complex)
waitGroups. However that would have not work for TestFetcherFactory.
As fetcherFactory.New()starts a goroutine inside it's body without
providing a way for clear termination. (We did not want to modify the
interface for this one, as we think the unclean termination does not
cause a problem in this case in production.)

fixes ethersphere#1109

**note**: alternative to https://github.com/ethersphere/go-ethereum/pull/1121