@frncmx 2019-01-17
Shall I also change `var RequestTimeout` as it currently does not cause a data race, but it uses the same change value and set back with defer logic in tests.