@janos 2019-01-17
I have run a simple benchmark and there is no significant overhead with additional Fetcher field searchTimeout on construction.

```
var benchmarkNewFetcherFetcher *Fetcher

func BenchmarkNewFetcher(b *testing.B) {
	requester := newMockRequester()
	addr := make([]byte, 32)
	var f *Fetcher
	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		f = NewFetcher(addr, requester.doRequest, true)
	}
	benchmarkNewFetcherFetcher = f
}
```

`go test -benchmem -run=none -cpu=1,2,4,8 github.com/ethereum/go-ethereum/swarm/network -bench BenchmarkNewFetcher -v -count=1`

This branch:

```
goos: darwin
goarch: amd64
pkg: github.com/ethereum/go-ethereum/swarm/network
BenchmarkNewFetcher     	 5000000	       235 ns/op	     272 B/op	       4 allocs/op
BenchmarkNewFetcher-2   	10000000	       165 ns/op	     272 B/op	       4 allocs/op
BenchmarkNewFetcher-4   	10000000	       176 ns/op	     272 B/op	       4 allocs/op
BenchmarkNewFetcher-8   	10000000	       163 ns/op	     272 B/op	       4 allocs/op
PASS
```

master:

```
goos: darwin
goarch: amd64
pkg: github.com/ethereum/go-ethereum/swarm/network
BenchmarkNewFetcher     	10000000	       222 ns/op	     272 B/op	       4 allocs/op
BenchmarkNewFetcher-2   	10000000	       160 ns/op	     272 B/op	       4 allocs/op
BenchmarkNewFetcher-4   	10000000	       158 ns/op	     272 B/op	       4 allocs/op
BenchmarkNewFetcher-8   	10000000	       162 ns/op	     272 B/op	       4 allocs/op
PASS
```

Since performance is not affected, contrary to my assumption. The question about correctness of the code remains. But I would go with the practical approach of accepting the simples solution, as it adds much less cognitive overhead.