@nonsense 2019-01-17
@janos I agree that this is not adding much overhead, but it just feels wrong to copy the timeout in the `fetcher`. Conceptually this parameter is static, so I'd rather we change the tests, and not add redundant fields.