@janos 2019-01-17
@nonsense the argument of redundancy is a strong one. I have no strong preference over additional field in struct or different, and possibly, more complex tests. I approved this PR as it is currently the simplest one that solves the problem without impacting production.