@frncmx 2019-01-17
@nonsense 
> We should probably refactor the tests, rather than store the timeout in every single fetcher.

We tried the fix from the test's side. However, the results were much more complex than this solution:
- approach A) https://github.com/ethersphere/go-ethereum/pull/1121
- approach B) waitGroups https://github.com/ethersphere/go-ethereum/pull/1121#pullrequestreview-193098010 - does not solve one case, as there is no way to wait for clean shutdown

Please, take a look at the solution ideas above. **Consider the alternatives complexity**. I think optimizing code for humans first is a good idea until we did not clearly identify a bottleneck.

Also, having a package level variable is global, mutable state for a package. That does not seem a good idea for a highly concurrent language. I think the variable belongs to the Fetcher object as it used by that specific instance in its worker loop, usually running in a separate goroutine.

Regarding storage redundancy: `searchTimeout` is `type Duration` and that is just an extra `int64` variable. Storing it might seem a lot, but let's consider a function call with an integer. That's also by value and an `int` gets copied on every function call.


Given the complexity of the other alternative solutions and that this gives no (/close to nothing) performance penalty I would really like to go with this solution. What do you think?