@nonsense 2019-01-17
I am still against the current solution. You just need a waitGroup for tests that modify `searchTimeout`, and not on the `searchTimeout` itself. I don't think this is complex at all.