@nonsense 2019-01-17
@janos having spoken with @frncmx it seems like we don't have functionality to know when the `go fetcher.run(...)` goroutine exits, which appears to be the root cause of these issues.

We should refactor this code, and add a timeout or cancellation, but for now I am okay with adding this temporary fix.